Javadoc and package rearrangement
