+++ src/main/java/org/cru/godtools/api/translations/GodToolsTranslationRetrieval.java
@@ -4,10 +4,8 @@ import com.google.common.base.Strings;
 import com.google.common.base.Throwables;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
-import org.ccci.util.xml.XmlDocumentStreamConverter;
 import org.cru.godtools.api.packages.utils.FileZipper;
-import org.cru.godtools.api.translations.config.Config;
-import org.cru.godtools.api.translations.contents.Content;
+import org.cru.godtools.api.translations.model.Content;
 import org.cru.godtools.api.translations.drafts.DraftUpdateJobScheduler;
 import org.cru.godtools.domain.GodToolsVersion;
 import org.cru.godtools.domain.GuavaHashGenerator;
@@ -23,7 +21,6 @@ import org.w3c.dom.Element;
 import javax.enterprise.inject.Default;
 import javax.inject.Inject;
 import javax.ws.rs.NotFoundException;
-import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
