+++ src/main/java/org/cru/godtools/api/translations/GodToolsTranslationService.java
@@ -4,7 +4,7 @@ import com.google.common.base.Optional;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 
-import org.cru.godtools.api.translations.config.Config;
+import org.cru.godtools.api.translations.model.Config;
 import org.cru.godtools.api.translations.drafts.DraftUpdateJobScheduler;
 import org.cru.godtools.api.cache.GodToolsCache;
 import org.cru.godtools.domain.GodToolsVersion;
