+++ src/main/java/org/cru/godtools/api/translations/model/Content.java
@@ -1,4 +1,4 @@
-package org.cru.godtools.api.translations.contents;
+package org.cru.godtools.api.translations.model;
 
 
 import com.google.common.collect.Sets;
@@ -11,6 +11,21 @@ import javax.xml.bind.annotation.XmlRootElement;
 import java.util.Collection;
 import java.util.Set;
 
+/**
+ * Represents a contents file that has 'meta' information about a God Tools resource.
+ *
+ * It contains a set of @see Resource s as XML elements or JSON objects..  Each resource represents a package or resource - a piece of content in the God Tools world.
+ * A resource could be "Knowing God Personally" for example.  There may be one or many Resources in the set.
+ *
+ * Resources are not nested inside a 'resources' element -- FIXME: uhh they currently are.
+ *
+ * Its XML representation looks like this example (some pages omitted):
+ *
+ * <?xml version="1.0" encoding="UTF-8"?>
+ *     <content>
+ *         <resource config="2c510fda-fb80-4c22-9792-195f36232b48.xml" icon="30adeee865dd2ff568b11715e9077ffbb851bb65.png" language="fr" name="Connaître Dieu Personnellement" package="kgp" status="live" version="1.1"/>
+ *     </content>
+ */
 @XmlRootElement
 public class Content
 {
